+++ packages/@expo/package-manager/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎉 New features
 
 - Support Bundler-managed CocoaPods installations ([#43605](https://github.com/expo/expo/pull/43605) by [@tiwari91](https://github.com/tiwari91), [@kitten](https://github.com/kitten))
+- Add Deno support with `DenoPackageManager` and `deno.lock` detection.
 
 ### 🐛 Bug fixes